notification-service/pyproject.toml
[project]
name = "notification-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.71",
    "fastapi>=0.135.1",
    "pika>=1.3.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
]
